requirements_test.txt
numpy
pytest
replicate
requests
Pillow